rename textsize file
